fixes to properly go to fastworkflow for command key mapping and fuzzy match can now be accessed through utils folder
